Introduce second z value for moving through depth during video clip.
Each scan point may have a second z value called z1 which is the value to
which the z-axis will move during the video clip.

For micro-5 this allows us to start with the focal plane near one membrane
and move through the culture during the clip in order to better the
C. elegans across the depth of the culture.