implement "just corners" scan; fix cell 4,5,6 offsets; reach as far as can
Oops! Got ahead ahead of myself and have three commits in one.

just corners:
Have a just_corners bool for scan generation that just hits the corners
that will be created with a full scan. For instance: if the full scan will
have 4 rows and 3 columns, it only sends the camera to (0,0), (2,0), (0,3),
and (2,3). Should be useful for checking locations before running full scan

fix 4,5,6 offsets:
The y values in the generate_six_wel_xy_corners for cells 4, 5, and 6 are
now corrected.

Only reach as far as you can:
We may be getting close to our maximum reach at the uppermost and
bottommost camera locations. If it can't quite reach, catch the trigonomety
value exception and set the camera arm to straight up or straight down
which is far as we can reach.