Handle various time value situations
If there is no 't' key, it is an intermediate scan point that has no video (image or clip)

t = 0 is a placeholder for taking an image instead of clip

t > 0: t is the integer duration of the video clip in seconds.
       It is also used to determine the speed of the second z move
       in instances where we are imaging while moving through the
       depth of the culture

If the t value is too small for the second z move to be able to achieve without exceeding the max speed, the duration is extended to match the max speed.