major rearrangement to make scans into classes
There is a base scan class which has an empty scan list, id string, and
scan building function that generically builds a scan that scans across
an arbitrary number of equally sized rectangles when given a list of
their top left corners.

The six well scan inherits from the base scan and has knowledge of the
size and arrangement of the six wells with respect to one another. Its
constructor takes the top left corner of the top left well and builds
the others based on the known geometry.

It also has a six well class that only includes the four corner scan
points of each well. It is intended for use checking the alignment of
the plate.