query camera info in ueye_camera constructor
Make "idscam info ..." systemm call and parse the data returned on stdout
to determine the sensor resolution. This info will be use to handle some
aspects of setting subsampling and binning values later.

For now, the idscam software still returns a zero even when it can't find
a camera with the given id, so we are checking for a stderr result to
determine failure. Will switch when idscam is updated.