Use argparse to handle user variables
Argparse is being used to handle both command line arguments and config file
parameters. The intent is that the config file parameters values are identified
by the developer or test engineer during testing and are unlikely to be altered
during use. An example would be the camera warm up time. It would probably be
determined heuristically during testing and isn't something that a user should
have to know about. The use (if not format) of the config variables is modeled
after the BioServe C config library. Command line values for these config
parameters supercede those loaded from file.

All other arguments are handled in the normal way and assumed to be set by the
user explicitly or the defaults are used.