Move generation of 6-well origins into function
It becomes member function of six_well_scan and the six_well_just_corners_scan
becomes child of six_well_scan.

Also beginning transition of language to "origins" for the corner of the well
because the "top-left corner of the top-left well" is unweildly and particularly
because it was confusing whether we were talking about a single corner or the
scan that includes the four corners of each well.