Add TODO for giving camera chance to warm up
If the ueye daemon has been off since computer start up, the camera will not
have downloaded firmware and initialized. Scancam can currently see that the
daemon is off and start it, but it does not give the camera a chance to
initialize before making a call to it and errors out.