Build scan areas positively from origin
The scan target areas were previously incremented positively in the Y direct
and negatively in the X direction.

The change redefines the target area origins as being the corner of the area
with the minimum X and Y coord values. It also defines each camera position
to increment positively in both X and Y as it scans over the target area.