Add 4 per well scan
This scan has two X and two Y positions. It only covers a third of the well
area, but each area is spaced apart but somewhat centrally. It is intended
to be used to check the min and max depths particularly for an over-filled
well whose membranes may be bulged.