Use microplate-like ids for wells in 6-well plate
Standard microplates have an enumeration system that uses a matrix scheme
for identifying the different wells. The long axis of the plate counts in
numbers, the short axis increments in letters. The top-left corner is therefore
'A1'. We use the same system for identifying the wells in a six well plate.

It is a step trickier to assign when creating the 6-well scan. Different
configurations of the scancam and orientations of the biocell could create
eight different combinations of 90 degree rotations and right or left handedness
of the scancam coordinate system. Diagonal rotations are not supported.

For the purposes of micro-5, there are only two orientations that will be used.
Both have the long axis of the plate in the Y direction. One is rotated 180
degrees from the other.

The update is that the choice between the two orientations is passed in when
creating the scan. The Module1 orientation is '1' and Module2 is '2'.

Now each point generated in the scan has a 'area-id' key with the appropriate
id value.