add scan building flag to always start with z0
A scan that has both z0 and z1 values performs a z-axis move while the video
clip is being recorded. When building scans, the points typically alternate
which value it starts the clip with and which it finishes with.

In some cases, like depth calibrations, we would prefer to always start and
finish in the same positions. This way, we can correlate the elapsed time
in the video clip with the position of the z-axis (depth). For instance, if
we start at 0.0mm and move to 7.0mm in 7 seconds then there is a one-to-one
relation between the elapsed time in seconds and the position in mm.

It is a flag passed in at scan construction and passed through to the scan
building function.