take running start on depth to make up for camera
When there is a z1 value in addition to the z0 value, we are going to record
video while walking through the depth of a culture by perfomring a z-axis move.

The 5 second delay when starting video with lights is starting to be
significant when compared to the likely 15-30 second video clips that we are
taking for micro-5. In order to account for the camera warm up delay, we back
up and take a "running start" on the z move. The timing is such that we are in
the desired starting z position when the camera starts to record video.