set all scans to current best guess
The flight scan is set to 20 seconds clip and no z values. It assumes that we
will add them in a calibration step.

The just-corners scan is set to 3 second clip at a central depth. Any shorter
than 3 seconds seems unweildly and we aren't particularly worried about focus
since we are just looking for the edges to check lateral alignment.

The central depth-calibrating scan reduces from four points per well to two.
(One might even be okay). It can no longer start at zero since we now have
to take the "running start" to allow for the camera warm-up. It takes 60
seconds to move from depth 1.0mm to 7.0mm. This has the convenient conversion
of 0.1 mm/sec which should be helpful when calibrating the depth.